UIAPPS-151 Move `endpoint`/`method` inside helper
We don't want to have three different things that have to be in sync.

The previous way these functions worked, we had to have all three of
these things line up with each other. If the `endpoint` wasn't the one
for `'PATCH'`, but we sent in `'PATCH'` anyway, it would break; if the
`endpoint` was for `'PATCH'`, but we didn't send the `appId`, it would
break; etc.

Rather than make sure these things stay in line, we take the actual
thing that decides the rest of the data: the `appId`. If it exists, the
`endpoint` should be `/api/v2/apps/{app_uuid}` and the `method` should
be `'PATCH'`. If it doesn't exist, the endpoint should be `/api/v2/apps`
and the `method` should be `'POST'`. The helper is now simpler to use,
and more correct from an implementation standpoint.