UIAPPS-93 Setup template for `--features` option
We copy most of the code in the `starter-kit` example.

For the most part what we're doing is taking the `starter-kit` example
and allowing App developers to choose only the parts they want.

The way this works is explained more in the `DESIGN.md` file. But for
the sake of the commit, we're generating whatever is given in the option
and only what is given in the option. We use handlebars as the
templating language to generate the the files, but only where necessary.

We'll want to expand on this more as we get more feedback. For now, this
is the first pass at having *something* that we can build on top of.