UIAPPS-93 Extract loggers to private methods
With these at the class-level, it should be easier to work with.

We want to extract methods from the `execute` method. That'll get a
little annoying to pass around information. We're moving to having the
loggers work with the state they already have. Should make future
changes easier to deal with.