UIAPPS-93 Explicitly include the `.gitignore`
We need this file as part of the template files.

Turns out that `yarn pack` and `npm pack` work slightly differently. For
`yarn pack`, it will include the `bin/template/.gitignore` file. For
`npm pack`, it will exclude the `bin/template/.gitignore` file. Since
that file helps when getting started with stuff, it's something we want
to include in the template.

We explicitly include the `bin/template/.gitignore` file to make sure it
shows up in the tarball.