UIAPPS-90 Start testing custom widget client
We're about to make changes, so we want to validate the behavior.

Even if this behavior isn't what we want, at least we want to make sure
it's visible somewhere. In particular, it doesn't seem right that we're
allowing the `updateOptions` method to make the request if the feature
isn't enabled. We'll address that, but we want to at least get the
current behavior tested.