UIAPPS-90 Check custom widget feature is enabled
We want to be consistent with our checking before making requests.

This seemed to be the only request made (or received) that didn't make
sure the feature was enabled. We change it so that we the feature has to
be enabled before sending the request.

It's likely that when all the pieces were moving together, the system
didn't allow this state to exist. But from the SDK's perspective, it was
inconsistent with what was going on in other feature clients. For that
sake alone, we make sure to validate before sending the request.