UIAPPS-90 Pull up validation to `DDFeatureClient`
We make it so receiving/sending a request does the validation.

Instead of making each subclass remember to do the validation, we pull
the similar logic up to the superclass of `DDFeatureClient`. In order to
mitigate us backsliding and forgetting to validate, we make the `client`
property private and instead provide some `protected` methods that do
the logic properly. This should hopefully make it so we don't have to
worry about remembering to validate in the future.

We also add a few tests to the `DDDashboardCogMenuClient` and
`DDWidgetContextMenuClient` test suites. We weren't checking that things
were validating before, but these tests should make sure we do now.