Remove a bunch of dead code
There was quite a bit of stuff that was hanging around.

We get rid of a couple of enums: `AuthStateStatus` and `Host`. It's not
clear when either of these was last used, but the latter is a bit out of
date given the state of the world now. We have Apps on multiple
datacenters now, so to have a single `PROD` doesn't really work.

We also remove the `MockLocalStorage` class. This seems to have last
been used when we supported secrets. That code has been removed for
quite some time, but we missed removing this class.

Since these things were exported as part of the public API, we make sure
to bump the version so we don't accidentally break someone using semver.