UIAPPS-90 Handle custom widget options requests
We also deprecate sending custom widget options requests.

The changeset details more of what we're doing, but the long and short
is that we move from the App developer being in control of dynamic
options to the App developer responding to requests for dynamic options.
This is more in line with how other dynamic behavior works and allows us
to provide a better experience in the end.

We've marked the old method as deprecated and will remove it at some
point. We also update the documentation to explain to App developers how
to respond to dynamic options.