UIAPPS-147 Make a hook for custom widget options
We want to make it a bit easier to deal with custom widget options.

As mentioned in the changeset, there's a non-trivial amount of work thatt
has to go into properly working with custom widget options. We make a
small hook that should make that easier to deal with.

Instead of having to work directly with the context and events, this
hook abstracts that away and gives a more declarative interface for
custom widget options. The results of using it in the example Apps seems
show a lot of the benefits. There's whole `useEffects` that don't really
need to be done now. And given that both of these examples didn't setup
the unsubscribe handlers properly, it seems like a net improvement at
least from how much App developers need to be concerned with.

Importantly, we only work with the actual options, and don't setup
anything for setting dynamic options. That feature is being deprecated
in https://github.com/DataDog/apps/pull/107, so it doesn't seem
worthwhile to add support for it to this hook.