UIAPPS-90 Don't use deprecated method in dog image
We update to the newer approach in the controller.

This is a real-world example of what the update might look like. Some
other updates might be a bit tricker, but for the most part we do all
the async stuff in the controller and let the custom widget focus on
being a custom widget.