UIAPPS-152 Make a hook for template variables
We want to make it a bit easier to work with template variables.

As mentioned in the changeset, there's a non-trivial amount of work that
has to go into properly working with template variables. We make a small
hook that should make that easier to deal with.

Getting the context and the handlers set up right is enough work on its
own, but App developers also have to know how to get the template
variables out of the context (e.g. they have to know that `name` is the
unique key). Hopefully, this hook makes things a bit easier.

The updated geomap example seems to indicate that things can be much
simpler, given that we dropped about 30 lines of boilerplate all for one
template variable.