UIAPPS-152 Listen to template change event as well
This commit epitomizes why a hook like this is useful.

We added the previous commit thinking we did entirely the correct thing,
but we completely missed the `DASHBOARD_TEMPLATE_VAR_CHANGE` event. If
we missed it, surely App developers are also missing it. And to that
point, even the example geomap App missed that event.

Creating a hook like this makes it so we only have to get it right once,
and everyone else can benefit from that work.