UIAPPS-147 Update versions
We want to make sure we keep these up to date.

As mentioned in the review comment:
https://github.com/DataDog/apps/pull/108#discussion_r803053815, we have
to manually update these, and it'd be nicer to just automatically have
them get updated whenever versions get released like the rest of the
packages.

We should try to figure this out sooner rather than later. It's an
annoyance that doesn't need to exist.