Start on a go example
This example shows how we can build an App in golang.

The general idea is: write go, compile it to wasm, run the wasm in the
browser. It works fairly well for something that doesn't need any
visual, like dashboard cog menu items or widget cog menu items.

We provide a small go package for interfacing with the SDK. It does the
brunt of the work, while the rest of the changes here are showing how to
actually pull the example together.