Make an example with the TypeScript API client
We can actually run it in the browser!

It's rough around the edges, but we get the ability to actually use the
TypeScript API client and all of its type information for each API it
supports.

We should consider extracting this out to either the SDK or its own
package, so that it's a bit easier for others to also use this API
client. But that's something we can think about in the future.